Fixing serialization issue, adding legacy prices for web

+++ api/src/main/scala/hr/com/blanka/apartments/common/DayMonth.scala
@@ -2,7 +2,7 @@ package hr.com.blanka.apartments.common
 
 import java.time.LocalDate
 
-case class DayMonth(day: Int, month: Long)
+case class DayMonth(day: Int, month: Int)
 object DayMonth {
   def apply(day: LocalDate): DayMonth =
     new DayMonth(day.getDayOfMonth, day.getMonthValue)
